+++ Tabby/src/Tabby/Scene/Components.h
@@ -314,6 +314,8 @@ struct Rigidbody2DComponent {
 };
 
 struct BoxCollider2DComponent {
+    std::pair<UUID, entt::entity> AtachedRigidbody2DEntity;
+
     glm::vec2 Offset = { 0.0f, 0.0f };
     glm::vec2 Size = { 0.5f, 0.5f };
     float angle = 0.0f;
@@ -351,6 +353,8 @@ struct BoxCollider2DComponent {
 };
 
 struct CircleCollider2DComponent {
+    std::pair<UUID, entt::entity> AtachedRigidbody2DEntity;
+
     glm::vec2 Offset = { 0.0f, 0.0f };
     float Radius = 0.5f;
 
@@ -387,7 +391,9 @@ struct CircleCollider2DComponent {
 };
 
 struct CapsuleCollider2DComponent {
-    glm::vec2 Offset = { 0.0f, 0.0f };
+    std::pair<UUID, entt::entity> AtachedRigidbody2DEntity;
+
+    glm::vec2 center1, center2;
     float Radius = 0.5f;
 
     // TODO(Yan): move into physics material in the future maybe
@@ -438,9 +444,7 @@ template <typename... Component>
 struct ComponentGroup { };
 
 using AllComponents = ComponentGroup<TransformComponent, SpriteRendererComponent,
-    CircleRendererComponent, CameraComponent, SoundComponent,
-    HierarchyNodeComponent, NativeScriptComponent,
-    Rigidbody2DComponent, BoxCollider2DComponent,
-    CircleCollider2DComponent, TextComponent>;
+    CircleRendererComponent, CameraComponent, SoundComponent, HierarchyNodeComponent, NativeScriptComponent,
+    Rigidbody2DComponent, BoxCollider2DComponent, CircleCollider2DComponent, CapsuleCollider2DComponent, TextComponent>;
 
 } // namespace Tabby