+++ changelog.md
@@ -1,4 +1,4 @@
 Version: 23-04-2024_1 Alpha
 Changes:
 - Physics2D rework.
-    - Swtiched to box2d 3.0 (aka box2c)
+    - Capsule collider added